[DOCS]Tugas Pertemuan 25 dan 26

+++ app/Http/Controllers/SliderController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Slider;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class SliderController extends Controller
 {
@@ -21,6 +22,14 @@ class SliderController extends Controller
 
     public function store(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'title' => 'required|string|min:5',
+            'caption' => 'required|string|min:10',
+            'image' => 'required|image|mimes:jpeg,png,jpg|max:2048',
+        ]);
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator->errors())->withInput();
+        }
         $imageName = time().'.'.$request->image->extension();
         Storage::putFileAs('public/slider', $request->file('image'), $imageName);
         $slider = Slider::create([
@@ -39,6 +48,14 @@ class SliderController extends Controller
 
     public function update(Request $request, $id)
     {
+        $validator = Validator::make($request->all(), [
+            'title' => 'required|string|min:5',
+            'caption' => 'required|string|min:10',
+            'image' => 'image|mimes:jpeg,png,jpg|max:2048',
+        ]);
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator->errors())->withInput();
+        }
         if ($request->hasFile('image')) {
             $old_image = Slider::find($id)->image;
             Storage::delete('public/slider/'.$old_image);
